Keep local time helper strongly typed

+++ mobile/lib/core/time/local_time_format.dart
@@ -5,8 +5,8 @@ TimeOfDay timeOfDayFromLocalTime(String value) {
   final hour = parts.isNotEmpty ? int.tryParse(parts[0]) ?? 0 : 0;
   final minute = parts.length > 1 ? int.tryParse(parts[1]) ?? 0 : 0;
   return TimeOfDay(
-    hour: hour.clamp(0, 23),
-    minute: minute.clamp(0, 59),
+    hour: hour.clamp(0, 23).toInt(),
+    minute: minute.clamp(0, 59).toInt(),
   );
 }
